move subcommands to own sub packages
